auto_tq_win/requirements.txt
requests
allure-pytest
opencv-python
psutil
json_tools
pyscreenshot
pillow
pyperclip
jsonpath
pyautogui
pynput